Add experimental Errbit support
The option to install Errbit is still experimental because we're still
not sure whether it's a good idea to run Errbit on the same server as
the application is running, and so it isn't enabled by default and there
isn't even a place to configure it. To enable it, either edit your hosts
file to add `errbit=true` or add it as an environment variable when
executing ansible.

Note executing `rails runner` in Errbit results in many warnings and
info message, and so we're using the `mongo` command to get the app
api_key, which we would normally obtain with `App.last.api_key`.