Test without world-readable temporary files
We used to have a problem on some servers because we were executing the
postgresql module as an unprivileged user and then in that module we
became the postgres user, which is also unprivileged. This resulted in
the postgres user not being able to read temporary files Ansible created
as the first user.

We used the `allow_world_readable_tmpfiles` option to fix this issue.
However, with recent Ansible versions we're getting lots of warnings:

```
[WARNING]: Using world-readable permissions for temporary files Ansible
needs to create when becoming an unprivileged user. This may be
insecure.
```

While in commit 790c47aa6 we decided in our case this is not a security
risk because we trust the client machine, the huge number of warnings
we're getting now generates too much noise during the installation.

We can avoid this issue by runnig the posgresql module as root. Quoting
the Ansible documentation [1]:

> Everything is fine if the module file is executed without using
> become, when the become_user is root, or (...)

[1] https://docs.ansible.com/ansible/2.10/user_guide/become.html#risks-of-becoming-an-unprivileged-user