Use systemd to start Puma
The `puma.service` file included here is an adaptation of the ERB
template provided by the capistrano3-puma gem. In order to transition
smoothly from the systemd configuration generated by this installer to
the systemd configuration provided by capistrano3-puma, the
`puma_service_unit_name` must be the same during both installation and
deployment.

Note that, now that Puma starts with systemd, we have to wait till it
creates a socket before continuing; otherwise the socket file won't be
there when we execute the next command, and that command will file.

Also note we have to configure the XDG_RUNTIME_DIR variable. Quoting the
ansible documentation [1]:

> For systemd to work with ‘user’, the executing user must have its own
> instance of dbus started and accessible (systemd requirement).
>
> The user dbus process is normally started during normal login, but not
> during the run of Ansible tasks. Otherwise you will probably get a
> 'Failed to connect to bus: no such file or directory' error.
>
> The user must have access, normally given via setting the
> XDG_RUNTIME_DIR variable.

We also need to check for systemd access because ansible doesn't do it
automatically [2].

[1] https://docs.ansible.com/ansible/latest/collections/ansible/builtin/systemd_service_module.html#parameter-scope
[2] See issue in 72674 https://github.com/ansible/ansible/issues/