Modify the Puma command so it works with Ansible
While this isn't needed when starting the Puma service on the server or
with Capistrano, we need the RVM script while using the installer.

The file is now different from the one generated by capistrano3-puma,
but that's alright because it will be overwritten when deploying with
Capistrano (as long as the `:puma_service_unit_name` is the same).