Add support for Debian Bookworm
Since Bookworm comes with PostgreSQL 15, we need to make the
`deploy_user` the database owner. Quoting the PostgreSQL 15 release
notes [1]:

> PostgreSQL 15 also revokes the CREATE permission from all users
> except a database owner from the public (or default) schema.

Now, we first create the user without any privileges, then create the
database with the new user as the database owner and finally grant
all privileges to the new user on the new database.

[1] https://www.postgresql.org/about/news/postgresql-15-released-2526/

Co-Authored-By: Senén Rodero <senenrodero@gmail.com>