Use the default Python version in Ubuntu workflows
We were having problems with the python-setup action when trying to use
Ubuntu 24.04 because it uses Python 3.12 instead of Python 3.8 and
because the way to install ansible with pip has changed over the years.

So we're now using the default version of Python and installing ansible
with apt, just like we do in Debian workflows.